requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
trafilatura
tavily-python
streamlit
requests
